ResultWriterHDF5, anim_concentrations: write neighbors and graph them
anim_concentrations.py --connections model.h5 | neato -Tsvg > model-conn.svg && firefox model-conn.svg

+++ python/anim_concentrations.py
@@ -1,7 +1,10 @@
 #!/usr/bin/python
+from __future__ import print_function, division
+
 import sys
 import os
 import glob
+import random
 import argparse
 import subprocess
 import numpy
@@ -13,6 +16,7 @@ from matplotlib import pyplot
 parser = argparse.ArgumentParser()
 parser.add_argument('file', type=tables.openFile)
 parser.add_argument('--save')
+parser.add_argument('--connections', action='store_true')
 
 def draw(sim, save):
     concs = sim.concentrations
@@ -48,10 +52,30 @@ def make_movie(save):
                  -ovc lavc -lavcopts vcodec=mpeg4 -oac copy -o'''.split()
     subprocess.check_call(command + [save, 'mf://*.png'.format(save)])
 
+def dottify(dst, connections, couplings):
+    print('digraph Connections {', file=dst)
+    print('\trankdir=LR;', file=dst)
+    print('\tsplines=true;', file=dst)
+    print('\tnode [color=blue,style=filled,fillcolor=lightblue];', file=dst)
+    for i in range(connections.shape[0]):
+        for j, coupl in zip(connections[i], couplings[i]):
+            if j < 0:
+                break
+            print('\t{} -> {} [penwidth={}];'.format(i, j, 10*coupl + random.random()), file=dst)
+    print('}', file=dst)
+
+def dot_connections(sim):
+    connections = sim.root.simulation.neighbors
+    couplings = sim.root.simulation.couplings
+    dottify(sys.stdout, connections, couplings)
+
 if __name__ == '__main__':
     opts = parser.parse_args()
-    draw(opts.file.root.simulation, opts.save)
-    if opts.save:
-        make_movie(opts.save)
-        for fname in glob.glob('{}-*.png'.format(save)):
-            os.unlink(fname)
+    if opts.connections:
+        dot_connections(opts.file)
+    else:
+        draw(opts.file.root.simulation, opts.save)
+        if opts.save:
+            make_movie(opts.save)
+            for fname in glob.glob('{}-*.png'.format(save)):
+                os.unlink(fname)
